Create README.md
The introduction of AutoMLP. We try to polish the tools here and figure out the sciential problem for MLP accelerated on hardware platform. 
FPGA accelerated is an ideal choice for real-time classification. 
If you need this tool and have any problem about it, please don't hesitate to contact us. We will share the paper about this tool to help understand MLP on FPGA and this work.